Add WebSockets support and interaction with the SIFIS-Home DHT
Added support for WebSockets for the interaction with the
SIFIS-Home DHT. This has been added to both the codebases
for Group OSCORE and EDHOC. Meaning that the Group OSCORE
Client and the EDHOC Initiator can now be triggered by
messages coming from the DHT.

To make this possible we have created a number of topics
to exchange messages to/from the DHT. In addition a Python
script was created that can send appropriate messages to
the DHT depending on which application is to be triggered.

This commit also contains a few other fixes and improvements,
for instance about handling of the ACE Token and updating of
the certificates that come with Californium.

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>