Client apps wait for Server hostnames to be resolvable before proceeding
The Group OSCORE Client/Server applications now wait for the hostnames
of the Authorization Server and Group Manager to become resolvable
before starting the Token retrieval and Group joining.

The EDHOC Client applications now wait for the EDHOC Server hostnames
to become resolvable before running EDHOC.

This is in addition to the Clients waiting for communication to the
Server side to actually work before proceeding.

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>